Refactor verify checkpoint CLI to cli package
