+++ packages/cli/src/index.ts
@@ -6,6 +6,7 @@ export * from './watch-contract';
 export * from './reset/watcher';
 export * from './reset/state';
 export * from './checkpoint/create';
+export * from './checkpoint/verify';
 export * from './inspect-cid';
 export * from './import-state';
 export * from './export-state';
